fix(osu_mbw_mr): float32 reduce for HCCL; honor WORLD_SIZE env in Makefile
- osu_mbw_mr.py: reduce the per-pair timing sum in float32 instead of
  float64. HCCL has no float64/double reduce ("HCCL reduce: Unsupported data
  type at::kDouble"), so the benchmark crashed on NPU while passing on
  CPU/gloo. A sum of per-rank seconds needs no double precision.
- Makefile: `export WORLD_SIZE ?= 2` (conditional) so an environment prefix is
  honored without `-e`, e.g. `WORLD_SIZE=8 make mbw-mr DEVICE=npu`. The plain
  `=` form shadowed the env value, silently pinning every ${WORLD_SIZE} target
  to 2.

+++ pytorch_hccl_tests/osu/p2p/osu_mbw_mr.py
@@ -104,7 +104,10 @@ def mbw_mr(args):
 
         # Sum sender times on rank 0 (always a sender). Every rank must take
         # part in the reduction, so non-senders contribute 0.
-        t_sum = torch.tensor(local_t_sec, dtype=torch.float64).to(device)
+        # float32 (not float64): HCCL does not support a float64/double reduce
+        # ("HCCL reduce: Unsupported data type at::kDouble"); a sum of per-rank
+        # seconds needs no double precision.
+        t_sum = torch.tensor(local_t_sec, dtype=torch.float32).to(device)
         dist.reduce(t_sum, 0, op=dist.ReduceOp.SUM)
 
         if rank == 0:
